requirements.txt
Flask
Flask-Cors
tensorflow==2.15.0
numpy
Pillow
gunicorn
